+++ src/com/hi_hobby/coupon/CouponView.java
@@ -1,4 +1,4 @@
-package com.hi_hobby;
+package com.hi_hobby.coupon;
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -11,7 +11,7 @@ import org.json.simple.JSONObject;
 import com.hi_hobby.action.Action;
 import com.hi_hobby.action.ActionInfo;
 
-public class CouponGetOk implements Action {
+public class CouponView implements Action {
 	@Override
 	public ActionInfo execute(HttpServletRequest req, HttpServletResponse resp) throws IOException {
 		req.setCharacterEncoding("UTF-8");
@@ -20,7 +20,7 @@ public class CouponGetOk implements Action {
 		JSONObject resultJSON = new JSONObject();
 		
 		String couponNum = req.getParameter("couponNum");
-		resultJSON.put("couponNum", couponDAO.resisterCoupon(couponNum));
+		resultJSON.put("couponNum", couponDAO.viewCoupon(couponNum));
 		
 		out.print(resultJSON.toJSONString());
 		out.close();
