+++ src/com/hi_hobby/domain/dao/AdminDAO.java
@@ -1,4 +1,4 @@
-package com.admin.hi_hobby.dao;
+package com.hi_hobby.domain.dao;
 
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
