+++ src/com/hi_hobby/domain/dao/CouponDAO.java
@@ -1,4 +1,4 @@
-package com.coupon.hi_hobby.dao;
+package com.hi_hobby.domain.dao;
 
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
