+++ src/com/hi_hobby/order/OrderFrontController.java
@@ -1,4 +1,4 @@
-package com.hi_hobby;
+package com.hi_hobby.order;
 
 import java.io.IOException;
 
@@ -9,7 +9,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.hi_hobby.action.ActionInfo;
 
-public class InquiryFrontController extends HttpServlet {
+public class OrderFrontController extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		doProcess(req, resp);
@@ -25,22 +25,14 @@ public class InquiryFrontController extends HttpServlet {
 		String command = requestURL.substring(requestURL.lastIndexOf("/") + 1);
 		ActionInfo actionInfo = null;
 		
-		if(command.equals("InquiryWriteOk.in")) {
-			actionInfo = new InquiryWriteOk().execute(req, resp);
+		if(command.equals("OrderCreate.or")) {
+			actionInfo = new OrderCreate().execute(req, resp);
 		}
-		else if(command.equals("InquiryWrite.in")) {
-			actionInfo = new ActionInfo();
-			actionInfo.setRedirect(true);
-			actionInfo.setPath(req.getContextPath() + "/csWrite.jsp");
+		else if(command.equals("OrderCancel.or")) {
+			actionInfo = new OrderCancel().execute(req, resp);
 		}
-		else if(command.equals("InquiryModify.in")) {
-			actionInfo = new InquiryModify().execute(req, resp);
-		}
-		else if(command.equals("InquiryDelete.in")) {
-			actionInfo = new InquiryDelete().execute(req, resp);
-		}
-		else if(command.equals("InquiryView.in")) {
-			actionInfo = new InquiryView().execute(req, resp);
+		else if(command.equals("OrderView.or")) {
+			actionInfo = new OrderView().execute(req, resp);
 		}
 		else {
 			// 404 일 때 출력할 에러 페이지 경로 작성
